🔊 (00:44:38) get params from a request:
the params of a request are values that go specified into the URL, after
the question mark "?" character. These are key, value properties
separated by equal "=" character, and each property is separated by an
ampersan "&" character. These values can be accessed by: request.GET and
django get these parameters in a "QueryDict" type but we have to parse it as a dictionary to use its values.

In this case I used both methods to add a properties, one directly into the url, and the another one by
the params variable into request.get() function