♓ (1:05:35) define a method on Product model:
By using the decorator @property in the models.py we may define a method
on a Django model that behaves like a read-only attribute. This means
that you can access the method like a regular attribute, without having
to call it explicitly as a method, i.e. we do not use the "()" at the
end of the call