🌙 (1:11:26) serialize a property that was not defined in the serializers:
this case is using what it is known as serializer method field.
<method>_<field>()
it gets its value by calling a method on the serializer class it is
attached to. It can be used to add any sort of data to the serialized
representation of your object.

method: this turn into values like <validate,get,create,delete,update>

field: the property I wanna perform action with.

the serializer method referred, accepts a single argument(in addition to
self) which is the object being serialized