🐲 (1:20:35) saves the objects when validating data:
the "serializer.save()" on views.py also creates an instance where it executes the get_my_discount() method inside the serializers.py.

The idea here is when we use a serializers.SerializerMethodField() on
the ProductSerializer (serializers.ModelSerializer), it's assumed that there's an instance
attached to it; because there is a model attached to it. But in this
case we don't need to have that, instead what we can do to hedge the
object is to return None when the object has no id property or the
object is not instance of the Product. This way we prevent Django from
throwing errors looking for the get_<field>()