📁 (1:25:00) Response data with 500 status:
every time the body does not includes the title property.

This happens because when we request the "/api" endpoint, it calls the
api_home view that uses ProductSerializer, then it validates that the
data is comming throug does match with the requirements of the
ProductSerializer first, second it validates the requirements of the
Product model, the product model requires the title field, which means
that every time the object comming throug have not the title property
included, it returns an 500 status code SERVER_ERROR, but we should not return 500 errors, because these are not catched errors.

A direct fix for this problem only is to conditionate the process to
tell the server what to return in this case, so we response with an
status code 400 BAD_REQUEST, and a short feedback to the user {"invalid":"not good data"}