🎬 (1:38:56) create a product despite the field is a type number:
this is something that I gladly amaze of serializers that with a few
lines of codes it validates and create the proper data.
Even if the data given is an string of numbers (like "26.99") works!
when the data given to property price is something else like "sdf" it
returns a warning