🆒 (3:04:16) add the REST_FRAMEWORK settings:
adding the REST_FRAMEWORK and DEFAULT_AUTO_FIELD allow us to remove the
authentication_classes from the ProductListCreateAPIView, due to these
two methods does not required? (not sure if that's what he mean it)
For these case is protection enough with just permissions_classes.

In this lecture Justin alse talked about Throttling, that bassiclly
means to limit the number of request to the server in a period of time,
more info: https://www.django-rest-framework.org/api-guide/throttling/#throttling