🎂 (3:18:00) use viewsets to create a use a set of endpoints /api/v2
rest_framework.viewSets is a class that provides a set of mehtods for handling HTTP
requests (GET,POST,PUT DELETE, etc.) for a specific type of data (e.g.
models or other data sources).

viewSets is designed to make it easy to create a CRUD interface for a
model or other data source.

viewSets provides a number of different methods for handling HTTP
requests, including "list","create", "retrieve", "update" and "destroy",
which correspond to the different CRUD operations.

rest_framework.routers.DefaultRouter is a class that allows you to
automatically generate URL routes and viewsets for a set of Django
models or views.

the DefaultRouter class automatically generates URL routes for the
standard CRUD operations for a given set of viewsets. It provides a
simple, standard way to map HTTP verbs to Django views and URLs, making
it easier to build a RESTful API.

the router.urls attribute in routers.py returns a list of URL patterns
that should be included in the Django urlpatterns list.

By using DefaultRouter, we can easily create a set of URLs for out API,,
and the viewsets provides the logic for handling the different HTTP
verbs and CRUD operations. This can help to reduce the amount of
boilerplate code required to build a RESTful API and make it easier to
build scalable,flexible APIs.