😮 (3:24:55) use as_view() to define the actions by its request method
Justin says at (3:21:03) that router.url does not give the granular
control of the URLs he personally likes to have, like declaring the
exactly string in the URL and what view is triggering when the endpoint
is called.
Anyway, We can get use to the pattern defined with router.url

Justin argumented that when you have multiple URLs defined into
urlpatterns list and you include to this list the router.urls, there
will not be way to identify which url are supported by router.urls because even in
urlpatterns is not that visible. In this case the urlpatterns prints in
console the following list:
[
  <URLResolver <URLPattern list> (admin:admin) 'admin/'>,
  <URLResolver <module 'api.urls' from '/REST/backend/api/urls.py'> (None:None) 'api/'>,
  <URLResolver <module 'products.urls' from '/REST/backend/products/urls.py'> (None:None) 'api/products/'>,
  <URLResolver <module 'cfehome.routers' from '/REST/backend/cfehome/routers.py'> (None:None) 'api/v2/'>
]

For that reason Justin kind of suggest to use the as_view() method to
declare the methods and what action (list,destroy, retrive,...), but he
still prefers to use declare them in views.py rather that routers