🌬️ (3:26:34) add and define url to the serializer:
I had an error making this commit it was saying:
```ImproperlyConfigured at /api/products/
Could not resolve URL for hyperlinked relationship using view name "product-detail". You may have failed to include the related model in your API, or incorrectly configured the `lookup_field` attribute on this field.```

that was only because I forgot to add the url field to the class
ProductSerializer, just below the my_discount field.