🟣 (3:34:47)Declare email to serializer but without display it
the EmailField has to be defined as write_only to prevent the error:
Got AttributeError when attempting to get a value for field `email` on serializer `ProductSerializer`.
The serializer field might be named incorrectly and not match any attribute or key on the `Product` instance.
Original exception text was: 'Product' object has no attribute 'email'.